Move CoverageCard rendering funcs to CoverageCard module
