+++ packages/slate-editor/src/modules/editor-v4-coverage/components/CoverageElement/CoverageCard/CoverageCard.tsx
@@ -11,7 +11,7 @@ import {
     getCoverageImageUrl,
     getCoverageTitle,
     hasOnlyFileAttachment,
-} from '../../../lib';
+} from './lib';
 
 import './CoverageBlock.scss';
