--- packages/slate-editor/src/modules/editor-v4-coverage/lib/getCoverageTitle.ts
@@ -1,13 +1,0 @@
-import type { Coverage } from '@prezly/sdk';
-
-export function getCoverageTitle(coverage: Coverage): string {
-    if (coverage.attachment_oembed && coverage.attachment_oembed.title) {
-        return coverage.attachment_oembed.title;
-    }
-
-    if (coverage.attachment) {
-        return coverage.attachment.filename;
-    }
-
-    return 'Untitled';
-}
