--- packages/slate-editor/src/modules/editor-v4-coverage/lib/hasOnlyFileAttachment.ts
@@ -1,13 +1,0 @@
-import type { Coverage } from '@prezly/sdk';
-
-export function hasOnlyFileAttachment(coverage: Coverage): boolean {
-    if (coverage.attachment_oembed && coverage.attachment_oembed.description) {
-        return false;
-    }
-
-    if (coverage.attachment) {
-        return true;
-    }
-
-    return false;
-}
